refactor discounts into a module
